pyproject.toml
[project]
name = "gevaarlijkestoffendb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fitz>=0.0.1.dev2",
    "ipython>=9.6.0",
    "langchain>=1.0.3",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.2",
    "openai>=2.6.1",
    "psycopg2-binary>=2.9.11",
    "pymupdf>=1.26.5",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
    "unstructured[local-inference,pdf]>=0.14.9",
    "pdfminer.six>=20221105",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.0.0",
]
